refactor: Consolidated binary-related modules under `binary` namespace
- Move `pe`, `elf`, and `macho` modules to the new `binary` namespace for improved modularity.
- Update imports across the codebase to reflect the structural changes.
- Adjust function visibility modifiers in `pe` for consistency (`pub(super)` to `pub(crate)`).
- Add `binary` submodules to the core library for streamlined access.
