+++ fireball/src/binary/elf/parser.rs
