+++ fireball/src/binary/macho/segment.rs
