+++ fireball/src/binary/macho/symbol.rs
