+++ fireball/src/binary/pe/asm.rs
