+++ fireball/src/binary/pe/fire.rs
