+++ fireball/src/binary/pe/fire/analyze_from_entry.rs
@@ -6,7 +6,7 @@ use crate::{
 use std::sync::Arc;
 
 impl Pe {
-    pub(super) fn _analyze_from_entry(&self) -> Result<Arc<Block>, DecompileError> {
+    pub(crate) fn _analyze_from_entry(&self) -> Result<Arc<Block>, DecompileError> {
         self.analyze_block(&self.entry)
     }
 }
