+++ fireball/src/binary/pe/fire/analyze_from_file_offset.rs
@@ -6,11 +6,11 @@ use crate::{
 use std::sync::Arc;
 
 impl Pe {
-    pub(super) fn _analyze_from_virtual_address(
+    pub(crate) fn _analyze_from_file_offset(
         &self,
         address: u64,
     ) -> Result<Arc<Block>, DecompileError> {
-        let address = Address::from_virtual_address(&self.sections, address);
+        let address = Address::from_file_offset(&self.sections, address);
         self.analyze_block(&address)
     }
 }
