+++ fireball/src/binary/pe/fmt.rs
